requirements.txt
Django>=5.0
django-environ>=0.11.2

//th
pymongo
python-dotenv